get it working for terminal accessible buffer
